==> microbe/analysis/2-Analysis.Rmd <==
---
title: "Analysis"
author: "fionarhuang"
date: "2020-04-23"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

```{r}
suppressPackageStartupMessages({
  library(TreeSummarizedExperiment)
  library(dplyr)
  library(treeclimbR)
  library(ggtree)
  library(TreeHeatmap)
  library(ggplot2)
  library(viridis)
  library(ggnewscale)
})
```

# Data aggregation

```{r}
load("output/Preparation.RData")
```


Data is stored as a `TreeSummarizedExperiment` object

```{r}
lse_phy <- TreeSummarizedExperiment(assays = list(count = new_count),
                                    rowData = new_tax,
                                    colData = samp_data,
                                    rowTree = new_phy,
                                    rowNodeLab = rownames(new_count))
```

Data on internal nodes are generated by data on leaves.
```{r}
all_node <- showNode(tree = new_phy, only.leaf = FALSE)
tse <- aggValue(x = lse_phy, rowLevel = all_node)
```


Data is split according to the three time points : `0 M`, `4 M` and `12 M`.
```{r}
gr <- c("0 M", "4 M", "12 M")
bse <- lapply(gr, FUN = function(x) {
  xx <- colData(tse)$period %in% x
  tse[, xx]
})
names(bse) <- gr
```

# run `edgeR` 

For data in each category, we compare the microbial abundance between two different delivery methods (vaginal and c-section) using `edgeR`.

```{r}

# build model
res <- lapply(bse, FUN = function(x) {
  runDA(TSE = x, feature_on_row = TRUE,
        assay = 1, 
        design_terms = c("born_method"))
})

# extract the result table
out <- lapply(res, FUN = function(x) {
  nodeResult(x, n = Inf)
  })


```

# run `treeclimbR`

There are two steps: propose candidates and evaluate candidates.
```{r}
# get candidates
cand <- lapply(out, FUN = function(x) {
  getCand(tree = new_phy, 
          score_data = x, 
          node_column = "node", 
          p_column = "PValue", 
          sign_column = "logFC")
})


# evaluate candidates
best <- lapply(seq_along(out), FUN = function(i) {
  out.i <- out[[i]]
  cand.i <- cand[[i]]
  evalCand(tree = new_phy, type = "single",
           levels = cand.i$candidate_list,
           limit_rej = 0.05,
           score_data = out.i,
           node_column = "node",
           p_column = "PValue",
           sign_column = "logFC",
          use_pseudo_leaf = FALSE)
}) 

names(best) <- names(cand)

```

# Results 


```{r}
# results of treeclimbR
df <- lapply(best, FUN = function(x) {
  topNodes(object = x, n = Inf, p_value = 0.05)
})
lapply(df, head)

# nodes identified
loc <- lapply(df, FUN = function(x){x$node})
lapply(loc, length)

# the number of descendant leaves of identified nodes
lapply(loc, FUN = function(x){
  xx <- findOS(tree = new_phy, node = x, only.leaf = TRUE, self.include = TRUE)
  ux <- unlist(xx)
  length(unique(ux))
  })
```


Here outputs results of treeclimbR for visualization in next step
```{r}
save(bse, loc, file = "output/Analysis.RData")
```
