==> microbe/analysis/3-Visualization.Rmd <==
---
title: "Visualization"
author: "fionarhuang"
date: "2020-04-23"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE, cache=FALSE}
knitr::opts_knit$set(echo = TRUE, message = FALSE, warning = FALSE, error=TRUE)
```

```{r}
suppressPackageStartupMessages({
  library(TreeSummarizedExperiment)
  library(dplyr)
  library(treeclimbR)
  library(ggtree)
  library(TreeHeatmap)
  library(ggplot2)
  library(viridis)
  library(ggnewscale)
})
```

Load data
```{r}
load("output/Analysis.RData")
```

# Plot tree
```{r}
new_phy <- rowTree(bse[[1]])

# tree
fig_0 <- ggtree(new_phy, layout = "fan", size = 0.5,
                branch.length = "none", 
                open.angle = 50) %>% 
  rotate_tree(140)

fig_0
```

# Color branches and nodes

Branches that are detected in at least one time point are colored. The detected
nodes in period `0 M` are labeled as red points.
```{r}
# tree: color branches that are detected in at least one period
uloc <- unique(unlist(loc))
ubr <- findOS(tree = new_phy, node = uloc, 
              only.leaf = FALSE, self.include = TRUE)
ubr <- unlist(ubr)

all_node <- showNode(tree = new_phy, only.leaf = FALSE)
df_br <- data.frame(node = all_node, Detected = "F") %>%
  mutate(Detected = ifelse(node %in% ubr, "Yes", "No"))


# tree (colored branch) + result
fig_0 <- fig_0 %<+% df_br + aes(color = Detected)+
    geom_point2(aes(subset = (node %in% loc[[1]])), 
                      color = "red", size = 2) +
  scale_color_manual(values = c("Yes" = "orange", "No" = "grey")) +
  guides(color = guide_legend(override.aes = list(size = 2)),
         fill = guide_legend(order = 1))


fig_0
```


# Black bars

Below add bars to indicate branches detected in different periods.

```{r}
# the alias labels of leaves on detected branches.
loc_leaf_alias <- lapply(loc, FUN = function(x) {
  lx <- findOS(tree = new_phy, node = x, 
               only.leaf = TRUE, self.include = TRUE)
  lx <- unlist(lx)
  transNode(tree = new_phy, node = lx, use.alias = TRUE)
})

# A binary matrix: Yes (DA) & No (non-DA)
leaf <- showNode(tree = new_phy, only.leaf = TRUE)
leaf_alias <- transNode(tree = new_phy, node = leaf, use.alias = TRUE)
mat_result <- matrix(NA, nrow = length(leaf), ncol = 3)
rownames(mat_result) <- leaf_alias
colnames(mat_result) <- c("0 M", "4 M", "12 M")

for (i in seq_along(loc_leaf_alias)) {
  mat_result[, i] <- ifelse(leaf_alias %in% loc_leaf_alias[[i]], "Yes", "No")
}

head(mat_result)
```

```{r}
# tree + bar (identified by treeclimbR)
col_sp <- colnames(mat_result)
names(col_sp) <- col_sp
fig_1 <- TreeHeatmap(tree = new_phy, 
                     tree_fig = fig_0, 
                     hm_data = mat_result, 
                     rel_width = 0.2,
                     tree_hm_gap = 0.3,
                     column_split = col_sp,
                     column_split_gap = 0.1,
                     show_colnames = TRUE,
                     colnames_position = "top",
                     colnames_size = 2, 
                     colnames_angle = 10, 
                     colnames_hjust = 0.5,
                     colnames_offset_y = 20,
                     colnames_offset_x = 0.25) +
  scale_fill_manual(values = c("Yes" = "black", 
                               "NO" = "white"), guide = FALSE) +
  new_scale_fill()

fig_1
```



# Add heatmaps

Average samples that are generated by randomly assigning samples that have the born method and at the same time point to five categories for each category. 

```{r}
# observed data: count (in cpm)
lse <- lapply(bse, FUN = function(x) {
  x[rowLinks(x)$isLeaf, ]
  })
count <- lapply(lse, FUN = function(x) {
  edgeR::cpm(assays(x)[[1]])
})

# Average samples: samples that have the born method and at the same time point are randomly assigned to five categories and an average sample is generated for each category.  
set.seed(1)
ind <- lapply(bse, FUN = function(x) {
  xx <- as.character(colData(x)$born_method)
  sx <- split(seq_along(xx), f = xx)
  lx <- lapply(sx, FUN = function(y) {
    y <- sample(y)
    nx <- length(y)/5
    split(y, ceiling(seq_along(y)/nx))
  })
})

# scale count: log transform + scale
ave_count <- ind
for (i in seq_along(ave_count)) {
  ind.i <- ind[[i]]
  lx <- lapply(seq_along(ind.i), FUN = function(x) {
    xx <- ind.i[[x]]
    xxx <- lapply(xx, FUN = function(y) {
    yy <- apply(count[[i]][, y], 1, mean)
    return(yy)
  })
  xc <- do.call(cbind, xxx)
  colnames(xc) <- paste(names(ind.i)[x], seq_len(ncol(xc)), sep = "_")
  return(xc)
  })
  ave_count[[i]] <- do.call(cbind, lx)
}

obs_ave <- lapply(seq_along(ave_count), FUN = function(x) {
  xx <- ave_count[[x]]
  lx <- log10(xx + 1) 
  ax <- apply(lx, 1, FUN = function(y) {
  #if (length(unique(y))) {y} else {scale(y)}
    ly <- length(unique(y))
    if (ly == 1) {
      ifelse(y > 0 , 1, 0)
    } else{
      sy <- scale(y)
      (sy - min(sy))/(max(sy) - min(sy))}
    })

  tx <- t(ax)
  rx <- rownames(xx)
  cx <- colnames(xx)
  nx <- names(ave_count)[x]
  colnames(tx) <- paste(nx, cx, sep = ":")
  return(tx)
 })
obs_ave <- do.call(cbind, obs_ave)

dim(obs_ave)
head(obs_ave)

```


```{r}
fig_2 <- fig_1
gr_C <- unique(gsub(pattern = ":.*", "", colnames(obs_ave)))
angle <- c(8, 5, 5)
for (i in seq_along(gr_C)) {
  gr_ci <- gr_C[i]
  sel_i <- grepl(gr_ci, colnames(obs_ave))
  mc <- obs_ave[, sel_i]
  head(mc)
  
  split_i <- ifelse(grepl(pattern = ":c_", colnames(mc)), 
                    "c-section", "vaginal")
  names(split_i) <- colnames(mc)
  
  split_lab_i <- split_i
  names(split_lab_i) <- split_i
  
  # 6 20
  fig_2 <- TreeHeatmap(tree = new_phy, 
              tree_fig = fig_2,  
              tree_hm_gap = 6 + 14*(i-1),
              column_split_gap = 0.2, 
              column_split = split_i,
              column_split_label = split_lab_i,
              split_label_size = 3,
              split_label_offset_y = 22 -(i) * 3 ,
              split_label_offset_x = -1,
              split_label_angle = angle[i],
              rel_width = 0.5,
              hm_data = mc,
              legend_title_hm = "Abundance") +
    guides(fill = guide_legend(order = 2))
  
}


fig_2 
```

# Label months


```{r}
# separate 0M, 4M, 12M with lines
hd <- getData(tree_hm = fig_2, type = "heatmap")
df_dec <- data.frame(x = c(max(hd$x) + 2, max(hd$x) - 12 , max(hd$x) - 27),
                      y1 = min(hd$y),
                      y2 = max(hd$y) + c(22, 18, 14),
                      label = c("12 M", "4 M", "0 M"), 
                      rotate = c(12, 8, 10),
                      ny = c(18, 18, 10))

fig_3 <- fig_2 + 
  new_scale_fill() +
  geom_segment(data = df_dec, 
               aes(x = x, y = y1, xend = x, yend = y2, group = label),
               linetype = "solid",
               inherit.aes = FALSE) +
  geom_label(data = df_dec, 
            aes(x = x, y = y2 + ny, label = label,
                angle = rotate, fill = label),
            #alpha = 0.4, 
            hjust = 1,
            show.legend = FALSE,
             inherit.aes = FALSE) +
  scale_fill_brewer(palette = "Set3") +
  new_scale_fill()


```

# Anotate genus

The genus of detected branches. We only show genus that has above five species (leaf nodes) detected.
```{r, warning=FALSE, message=FALSE}
# genus: detected
df_genus <- lapply(seq_along(loc), FUN = function(i) {
  bse.i <- bse[[i]]
  loc.i <- loc[[i]]
  des.i <- findOS(tree = new_phy, node = loc.i, 
                  only.leaf = TRUE, self.include = TRUE)
  des.i <- unlist(des.i)
  
  ri <- rowLinks(bse.i)$nodeNum %in% des.i
  
  bse.ri <- bse.i[ri, ]
  data.frame(Genus = as.character(rowData(bse.ri)$Genus),
             node = rowLinks(bse.ri)$nodeNum, 
             period = names(loc)[i])
  
})
df_genus <- do.call(rbind, df_genus) %>%
  group_by(Genus, period) %>%
  mutate(n = n()) 
dim(df_genus)

# genus: detected & above 5
df_genus_4 <- df_genus %>%
  filter(n > 4) %>%
  mutate(alias = transNode(tree = new_phy, 
                           node = node, use.alias = TRUE)) %>%
  ungroup() %>%
  select(Genus, alias) %>%
  distinct()

# the matrix is used later to label heatmap
mat_genus <- cbind(as.character(df_genus_4$Genus))
colnames(mat_genus) <- "Genus"
rownames(mat_genus) <- df_genus_4$alias

```

```{r, warning=FALSE, message=FALSE, fig.width=8, fig.height=7}
fig_4 <- TreeHeatmap(tree = new_phy, 
              tree_fig = fig_3,  
              tree_hm_gap = 50,
              rel_width = 0.1,
              hm_data = mat_genus,
              legend_title_hm = "Genus") +
  scale_fill_brewer(palette = "Paired") +
  guides(fill = guide_legend(order = 3))+
  theme(
    aspect.ratio = 1,
    legend.position = c(0.15, 0.66),
    legend.background = element_rect(fill = NA),
    plot.title = element_text(hjust = 0.5),
    legend.title = element_text(size = 9),
    legend.text = element_text(size = 7),
    legend.key.size = unit(0.4, "cm"),
    legend.spacing.x = unit(0.2, "mm"),
    legend.margin = margin(t = 0, b = 0, r = 5, l = 0),
    legend.box.margin=margin(t = -5, b = -5, r = 0, l = -60),
    plot.margin = margin(t = -50, b = -78, r = -200, l = -150)
  )

fig_4



```

```{r warning=FALSE, message=FALSE}
ggsave("output/baby_born.eps", fig_4, 
       units = "in", width = 8, height = 7,
       dpi = 300)
```
